make loading spinner standalone, move QuickAccessButton into Buttons
