+++ src/Components/src/Composite/Widgets/FilePickerWidget.fs
@@ -8,6 +8,7 @@ open Swate.Components.JsBindings
 open Swate.Components.Shared
 open Swate.Components.Primitive
 open Swate.Components.Primitive.Buttons
+open Swate.Components.Primitive.LoadingSpinner
 open Swate.Components.Composite.AnnotationTable.Context
 
 /// This context is designed to be used only internally in this file.
@@ -454,7 +455,7 @@ type FilePickerWidget =
             prop.className "swt:flex swt:flex-col swt:gap-2 swt:min-w-sm"
             prop.children [
                 if isLoading then
-                    Buttons.LoadingSpinner("Loading Paths...", DaisyuiSize.LG)
+                    LoadingSpinner.LoadingSpinner("Loading Paths...", DaisyuiSize.LG)
                 else if hasPaths then
                     let canInsert = hasActiveTableView && selectedCells.IsSome && hasPaths
