+++ src/Components/src/Primitive/Buttons/Buttons.fs
@@ -1,8 +1,10 @@
 namespace Swate.Components.Primitive.Buttons
 
 open Feliz
+open Fable.Core
 open Swate.Components.Primitive
 
+[<Erase; Mangle(false)>]
 type Buttons =
 
     [<ReactComponent>]
@@ -114,33 +116,40 @@ type Buttons =
         ]
 
     [<ReactComponent>]
-    static member LoadingSpinner(?text: string, ?size: DaisyuiSize, ?color: DaisyuiColors) =
-        Html.span [
-            prop.className "swt:flex swt:flex-col swt:items-center swt:gap-2 swt:py-10"
-            prop.children [
-                Html.div [
-                    prop.className [
-                        "swt:loading"
-                        match size with
-                        | Some(DaisyuiSize.XS) -> $"swt:loading-xs"
-                        | Some(DaisyuiSize.SM) -> $"swt:loading-sm"
-                        | Some(DaisyuiSize.MD) -> $"swt:loading-md"
-                        | Some(DaisyuiSize.LG) -> $"swt:loading-lg"
-                        | Some(DaisyuiSize.XL) -> $"swt:loading-xl"
-                        | None -> ()
-                        match color with
-                        | Some DaisyuiColors.Primary -> "swt:loading-primary"
-                        | Some DaisyuiColors.Secondary -> "swt:loading-secondary"
-                        | Some DaisyuiColors.Accent -> "swt:loading-accent"
-                        | Some DaisyuiColors.Warning -> "swt:loading-warning"
-                        | Some DaisyuiColors.Error -> "swt:loading-error"
-                        | Some DaisyuiColors.Info -> "swt:loading-info"
-                        | Some DaisyuiColors.Success -> "swt:loading-success"
-                        | None -> ()
-                    ]
-                ]
-                match text with
-                | Some t -> Html.span [ prop.text t ]
-                | None -> Html.none
+    static member QuickAccessButton
+        (
+            children: ReactElement,
+            title: string,
+            onclick: Browser.Types.MouseEvent -> unit,
+            ?isDisabled: bool,
+            ?props: IReactProperty seq,
+            ?color: DaisyuiColors,
+            ?classes: string
+        ) =
+        let isDisabled = defaultArg isDisabled false
+
+        Html.button [
+            prop.className [
+                "swt:btn swt:btn-ghost swt:btn-square swt:btn-transparent swt:bg-transparent swt:border-none swt:shadow-none"
+
+                match color with
+                | Some DaisyuiColors.Primary -> "swt:hover:text-primary!"
+                | Some DaisyuiColors.Secondary -> "swt:hover:text-secondary!"
+                | Some DaisyuiColors.Accent -> "swt:hover:text-accent!"
+                | Some DaisyuiColors.Error -> "swt:hover:text-error!"
+                | Some DaisyuiColors.Info -> "swt:hover:text-info!"
+                | Some DaisyuiColors.Success -> "swt:hover:text-success!"
+                | Some DaisyuiColors.Warning -> "swt:hover:text-warning!"
+                | None -> "swt:hover:text-primary!"
+
+                if classes.IsSome then
+                    classes.Value
             ]
+            prop.tabIndex (if isDisabled then -1 else 0)
+            prop.title title
+            prop.disabled isDisabled
+            prop.onClick onclick
+            if props.IsSome then
+                yield! props.Value
+            prop.children children
         ]
