+++ src/Components/src/Swate.Components.fsproj
@@ -93,7 +93,7 @@
     <Compile Include="Primitive\Types.fs" />
     <Compile Include="Primitive\Icons.fs" />
     <Compile Include="Primitive\Buttons\Buttons.fs" />
-    <Compile Include="Primitive\QuickAccessButton\QuickAccessButton.fs" />
+    <Compile Include="Primitive\LoadingSpinner\LoadingSpinner.fs" />
     <Compile Include="Primitive\Popover\Context.fs" />
     <Compile Include="Primitive\Popover\Popover.fs" />
     <Compile Include="Primitive\Select\Types.fs" />
